+++ Test/plot_error_test.py
@@ -18,14 +18,14 @@ Actual
 >>> ax = cm.plot()
 Traceback (most recent call last):
     ...
-pycm.pycm_error.pycmPlotError: Failed to import matplotlib module. Please install it using: `pip install matplotlib`.
+pycm.errors.pycmPlotError: Failed to import matplotlib module. Please install it using: `pip install matplotlib`.
 >>> ax = cm.plot(plot_lib='seaborn')
 Traceback (most recent call last):
     ...
-pycm.pycm_error.pycmPlotError: Failed to import matplotlib module. Please install it using: `pip install matplotlib`.
+pycm.errors.pycmPlotError: Failed to import matplotlib module. Please install it using: `pip install matplotlib`.
 >>> crv = Curve(actual_vector = np.array([1, 1, 2, 2]), probs = np.array([[0.1, 0.9], [0.4, 0.6], [0.35, 0.65], [0.8, 0.2]]), classes=[2, 1])
 >>> ax = crv.plot(classes=[1])
 Traceback (most recent call last):
     ...
-pycm.pycm_error.pycmPlotError: Failed to import matplotlib module. Please install it using: `pip install matplotlib`.
+pycm.errors.pycmPlotError: Failed to import matplotlib module. Please install it using: `pip install matplotlib`.
 """