+++ pycm/utils.py
@@ -6,7 +6,7 @@ import math
 import numpy
 import re
 from .params import *
-from .pycm_error import pycmMatrixError
+from .errors import pycmMatrixError
 from warnings import warn
 from functools import wraps